Add Generator service with recursive matching for vowels that are to be upcased
This implementation takes the natural approach, as if the substitution
is done by a human. As we progress left-to-right, character by
character, any substitutions done in the string earlier (to the left of
the current character) will affect the matches later on (to the right).
This is achieved by recursion in the block that gsub yields to.

However, with this approach, the second example given in the original
description (“The quick brown fox…”) will have 4 upper case vowels
after the conversions, not 3 as stated in the example checksum. I am
not sure if this is intentional or perhaps a typo in the example
checksum.